+++ src/wrappers/withDragHandles.js
@@ -77,6 +77,8 @@ export function dragHandleZone(node, options) {
             updateZone();
         },
         destroy: () => {
+            isItemsDragDisabled.set(true);
+            zone.destroy();
             node.removeEventListener("consider", consider);
             node.removeEventListener("finalize", finalize);
             isItemsDragDisabled.unsubscribe(updateZone);